[KT] Move internal headers into internal folder
Remove unused on_wg header
