+++ include/oneapi/dpl/experimental/kt/internal/esimd_radix_sort_dispatchers.h
@@ -10,13 +10,13 @@
 #ifndef _ONEDPL_KT_ESIMD_RADIX_SORT_DISPATCHERS_H
 #define _ONEDPL_KT_ESIMD_RADIX_SORT_DISPATCHERS_H
 
-#include "kernel_param.h"
+#include "../kernel_param.h"
 #include "esimd_radix_sort_utils.h"
 #include "esimd_radix_sort_submitters.h"
 #include "esimd_radix_sort_kernels.h"
 
-#include "../../pstl/hetero/dpcpp/parallel_backend_sycl_utils.h"
-#include "../../pstl/hetero/dpcpp/utils_ranges_sycl.h"
+#include "../../../pstl/hetero/dpcpp/parallel_backend_sycl_utils.h"
+#include "../../../pstl/hetero/dpcpp/utils_ranges_sycl.h"
 
 #if __has_include(<sycl/sycl.hpp>)
 #    include <sycl/sycl.hpp>
